Renamed new component files for easier identification
