+++ web/react/components/team_signup_allowed_domains_page.jsx
@@ -4,7 +4,7 @@
 var client = require('../utils/client.jsx');
 
 module.exports = React.createClass({
-    displayName: 'AllowedDomainsPage',
+    displayName: 'TeamSignupAllowedDomainsPage',
     propTypes: {
         state: React.PropTypes.object,
         updateParent: React.PropTypes.func
