+++ web/react/components/team_signup_password_page.jsx
@@ -5,7 +5,7 @@ var utils = require('../utils/utils.jsx');
 var client = require('../utils/client.jsx');
 
 module.exports = React.createClass({
-    displayName: 'PasswordPage',
+    displayName: 'TeamSignupPasswordPage',
     propTypes: {
         state: React.PropTypes.object,
         updateParent: React.PropTypes.func
