+++ web/react/components/team_signup_username_page.jsx
@@ -5,7 +5,7 @@ var utils = require('../utils/utils.jsx');
 var client = require('../utils/client.jsx');
 
 module.exports = React.createClass({
-    displayName: 'UsernamePage',
+    displayName: 'TeamSignupUsernamePage',
     propTypes: {
         state: React.PropTypes.object,
         updateParent: React.PropTypes.func
